requirements.txt
openai>=1.44.0
python-dotenv>=1.0.1
websockets>=12.0
httpx>=0.27.0
pydantic>=2.8.2
watchfiles>=0.21.0
